docs:  Make the generated docs visible.
There is still a lot of improvements that could be made to make the
generated docs better organized.  However, before this change we were
generating the docs but they weren't in the doc tree so you couldn't
actually browse to them.  So we fix that first so that we can start
getting feedback about how to organize them.